Refactor and clean up code across multiple modules
- Removed unnecessary blank lines and comments across the library.
- Updated `BaseAlgorithm` to allow for more flexible learning parameters and improved session handling.
- Enhanced `ccdf` and `confidence_interval` functions in `general.py` for better clarity and consistency.
- Updated logging functionality in `logger.py` to include unique identifiers for log directories.
- Refined network initialization in `network.py` by removing type ignores.
- Improved entropy annealing behavior in `bc.py` for better clarity and consistency.
- Adjusted Reinforce algorithm parameters in `reinforce.py` for better control over training.
- Enhanced `End2EndActor` class in `end2end.py` by removing type ignores.
- Improved `PlackettLucePolicy` in `perm.py` for better readability and added a new `WeightEstimationModel` class.
- Enhanced `TabularPreprocessor` in `preprocessor.py` to handle remaining actions more effectively.
- Updated protocols in `protocols.py` to allow for more flexible policy definitions.

+++ cpscheduler/gym/wrappers/obs_wrappers.py
@@ -25,9 +25,8 @@ def reshape_space(space: S, shape: tuple[int, ...]) -> S:
     "Reshape the space to the given shape."
     if isinstance(space, Box):
         return Box(
-            low=space.low.reshape(shape),  # type: ignore
-            high=space.high.reshape(shape),  # type: ignore
-            dtype=space.dtype,  # type: ignore
+            low=space.low.reshape(shape),
+            high=space.high.reshape(shape),
         )
 
     if isinstance(space, Dict):
